chore: preserve HorizonUI.Controls namespace
